cleanedup skunk attack and walk
