+++ Assets/Animations/SkunkMovement1.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: SkunkMovement
+  m_Name: SkunkMovement1
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
