+++ Assets/Animations/SkunkMovement4.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f6028d264d10f5479718f6a243520ea
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
