+++ Assets/Imported/skunk_leg.fbx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c875f9300d874e3b4fa044dbb1310a8a8882de1ed81bc5030ffad3b8ea4726b8
+size 16684
