+++ Assets/Materials/BoundaryMaterial.mat
@@ -74,5 +74,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 0
     m_Colors:
-    - _Color: {r: 0, g: 1, b: 0.052033663, a: 0}
+    - _Color: {r: 0, g: 1, b: 0.052033663, a: 0.40392157}
     - _EmissionColor: {r: 0, g: 1, b: 0.16476631, a: 1}
